SProperty fixes...
- SProperties now work again
- Add SAssetProperty
- Add missing value control for some SProperty types in SceneHierarchy
panel
- Remove char type
- Add GetCopy
- Fix HeaderTool bug

[no ci]

+++ SaturnHeaderTool/src/SaturnHeaderTool/Base/Core.h
@@ -48,9 +48,9 @@ namespace Saturn {
 		std::string ClassName;
 		std::string BaseClass;
 		uint64_t ClassFlags = 0 /* SClassFlags::None */;
-		uint32_t LineNumberForGeneratedBody = 0;
+		uint64_t LineNumberForGeneratedBody = 0;
 
-		std::map<uint32_t, SProperty> Properties;
+		std::map<uint64_t, SProperty> Properties;
 	};
 
 	struct HeaderToolCommand
